ajustes na GUI
