+++ common/metas.ts
@@ -1,0 +1,10 @@
+export class Metas {
+    meta: String [] = [];
+    peso: String = '';
+  
+    copyFrom(from: Metas): void {
+      this.meta = from.meta;
+      this.peso = from.peso;
+    }
+  }
+
