+++ ta-gui/src/app/turmas/turmas.component.ts
@@ -11,6 +11,7 @@ export class TurmasComponent implements OnInit {
   turmas: Turma[] = [];
   turma: Turma = new Turma();
   turmaMetas: string = '';
+  //turmaPeso: string = '';
   turmaRepetida: boolean = false;
   turmaEditar: Turma = new Turma();
   turmaEditarMetas: string = '';
@@ -103,7 +104,7 @@ export class TurmasComponent implements OnInit {
       let tempTurma = new Turma()
       tempTurma.descricao = turma.descricao;
       tempTurma.metas = turma.metas;
-      tempTurma.peso = turma.peso;
+     // tempTurma.peso = turma.peso;
       tempTurma.vagas = turma.vagas;
       this.turmasService.criar(tempTurma)
       .subscribe(
@@ -112,6 +113,7 @@ export class TurmasComponent implements OnInit {
             this.turmas.push(turma);
             this.turma = new Turma();
             this.turmaMetas = '';
+           // this.turmaPeso = '';
           } else {
             this.turmaRepetida = true;
           }
@@ -131,6 +133,7 @@ export class TurmasComponent implements OnInit {
             this.turmas.push(turma);
             this.turma = new Turma();
             this.turmaMetas = '';
+           // this.turmaPeso = '';
           } else {
             this.turmaRepetida = true;
           }
