+++ ta-server/package.json
@@ -5,12 +5,13 @@
   "main": "ta-server.js",
   "scripts": {
     "test": "npm run tsc && jasmine --config=spec/support/jasmine.json",
-    "tsc": "C:/Users/dinho/Documents/FACULDADE/ESS/ProjetoFinal/teaching-assistant/ta-server/node_modules/.bin/tsc",
+    "tsc": "C:/Users/acs/AppData/Roaming/npm/tsc",
     "start": "npm run tsc && node compiledCode/ta-server/ta-server.js"
   },
   "author": "Paulo Borba",
   "license": "MIT",
   "dependencies": {
+    "@angular/cli": "^11.2.11",
     "@types/body-parser": "^1.19.0",
     "@types/cors": "^2.8.8",
     "@types/express": "^4.17.8",
